Started working on inference/onnx. Onnxruntine not available for 3.10, annoying.

+++ src/dataset/data_transformations.py
@@ -40,7 +40,7 @@ def to_tensor():
 
         imgs = imgs.transpose((0, 3, 1, 2))
         # labels = rearrange(labels, "b w h c m -> b m c w h")
-        return torch.from_numpy(imgs).float().to(device), torch.from_numpy(labels).to(device)
+        return torch.from_numpy(imgs).to(device), torch.from_numpy(labels).to(device)
     return to_tensor_fn
